+++ Huobi.Net/Clients/SpotApi/HuobiRestClientSpotApiTrading.cs
@@ -14,6 +14,7 @@ using Huobi.Net.Objects.Models;
 using CryptoExchange.Net.Converters;
 using Huobi.Net.Interfaces.Clients.SpotApi;
 using CryptoExchange.Net.CommonObjects;
+using Huobi.Net.ExtensionMethods;
 
 namespace Huobi.Net.Clients.SpotApi
 {
