+++ Huobi.Net/Huobi.Net.xml
@@ -2202,6 +2202,17 @@
             Withdraw
             </summary>
         </member>
+        <member name="T:Huobi.Net.ExtensionMethods.HuobiExtensionMethods">
+            <summary>
+            Extension methods specific to using the Huobi API
+            </summary>
+        </member>
+        <member name="M:Huobi.Net.ExtensionMethods.HuobiExtensionMethods.ValidateHuobiSymbol(System.String)">
+            <summary>
+            Validate the string is a valid Huobi symbol.
+            </summary>
+            <param name="symbolString">string to validate</param>
+        </member>
         <member name="T:Huobi.Net.HuobiEnvironment">
             <summary>
             Huobi environments
@@ -2242,27 +2253,6 @@
             <param name="usdtMarginSwapRestAddress"></param>
             <param name="usdtMarginSwapSocketAddress"></param>
             <returns></returns>
-        </member>
-        <member name="T:Huobi.Net.HuobiHelpers">
-            <summary>
-            Helpers for Huobi
-            </summary>
-        </member>
-        <member name="M:Huobi.Net.HuobiHelpers.AddHuobi(Microsoft.Extensions.DependencyInjection.IServiceCollection,System.Action{Huobi.Net.Objects.Options.HuobiRestOptions},System.Action{Huobi.Net.Objects.Options.HuobiSocketOptions},System.Nullable{Microsoft.Extensions.DependencyInjection.ServiceLifetime})">
-            <summary>
-            Add the IHuobiClient and IHuobiSocketClient to the sevice collection so they can be injected
-            </summary>
-            <param name="services">The service collection</param>
-            <param name="defaultRestOptionsDelegate">Set default options for the rest client</param>
-            <param name="defaultSocketOptionsDelegate">Set default options for the socket client</param>
-            <param name="socketClientLifeTime">The lifetime of the IHuobiSocketClient for the service collection. Defaults to Singleton.</param>
-            <returns></returns>
-        </member>
-        <member name="M:Huobi.Net.HuobiHelpers.ValidateHuobiSymbol(System.String)">
-            <summary>
-            Validate the string is a valid Huobi symbol.
-            </summary>
-            <param name="symbolString">string to validate</param>
         </member>
         <member name="T:Huobi.Net.Interfaces.Clients.IHuobiRestClient">
             <summary>
@@ -9660,18 +9650,33 @@
             Extensions for the ICryptoRestClient and ICryptoSocketClient interfaces
             </summary>
         </member>
-        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Huobi(CryptoExchange.Net.Clients.ICryptoRestClient)">
+        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Huobi(CryptoExchange.Net.Interfaces.ICryptoRestClient)">
             <summary>
             Get the Huobi REST Api client
             </summary>
             <param name="baseClient"></param>
             <returns></returns>
         </member>
-        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Huobi(CryptoExchange.Net.Clients.ICryptoSocketClient)">
+        <member name="M:CryptoExchange.Net.Clients.CryptoClientExtensions.Huobi(CryptoExchange.Net.Interfaces.ICryptoSocketClient)">
             <summary>
             Get the Huobi Websocket Api client
             </summary>
             <param name="baseClient"></param>
+            <returns></returns>
+        </member>
+        <member name="T:Microsoft.Extensions.DependencyInjection.ServiceCollectionExtensions">
+            <summary>
+            Extensions for DI
+            </summary>
+        </member>
+        <member name="M:Microsoft.Extensions.DependencyInjection.ServiceCollectionExtensions.AddHuobi(Microsoft.Extensions.DependencyInjection.IServiceCollection,System.Action{Huobi.Net.Objects.Options.HuobiRestOptions},System.Action{Huobi.Net.Objects.Options.HuobiSocketOptions},System.Nullable{Microsoft.Extensions.DependencyInjection.ServiceLifetime})">
+            <summary>
+            Add the IHuobiClient and IHuobiSocketClient to the sevice collection so they can be injected
+            </summary>
+            <param name="services">The service collection</param>
+            <param name="defaultRestOptionsDelegate">Set default options for the rest client</param>
+            <param name="defaultSocketOptionsDelegate">Set default options for the socket client</param>
+            <param name="socketClientLifeTime">The lifetime of the IHuobiSocketClient for the service collection. Defaults to Singleton.</param>
             <returns></returns>
         </member>
     </members>
